requirements.txt
tensorflow==1.7
tensorflow-gpu==1.7
scikit-learn
opencv-python
h5py
Pillow
requests
psutil
aniso8601==1.2.0
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.3
boto3==1.4.4
botocore==1.5.14
click==6.7
colorama==0.3.7
cycler==0.10.0
decorator==4.0.11
docutils==0.13.1
enum34==1.1.6
Flask==0.12
Flask-RESTful==0.3.5
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.0.5
gevent==1.2.1
greenlet==0.4.12
happybase==1.0.0
ipython==5.2.2
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.9.5
jmespath==0.9.1
MarkupSafe==1.0
matplotlib==2.0.0
mock==2.0.0
numpy==1.12.0
olefile==0.44
packaging==16.8
pathlib2==2.2.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
ply==3.10
prompt-toolkit==1.0.9
protobuf==3.1.0
ptyprocess==0.5.1
pyasn1==0.2.2
Pygments==2.1.3
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.10
scandir==1.4
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
subprocess32==3.2.7
tfdeploy==0.3.3
thriftpy==0.3.9
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.12
